backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic>=2.10,<3
pydantic-settings>=2.6,<3

celery==5.4.0
redis==5.0.7
flower==2.0.1

langgraph==0.1.19
langchain-groq==0.1.6
langchain-community==0.2.6
langchain-core>=0.2.22,<0.3

chromadb==0.5.3
faiss-cpu>=1.9.0
sentence-transformers==3.0.1

pytesseract==0.3.10
Pillow==10.4.0
pdf2image==1.17.0
PyPDF2==3.0.1
python-multipart==0.0.9

httpx==0.27.0
python-dotenv==1.0.1
numpy
